removed unused function and fixed delete method.
the function `_fetch_live_server` was implemented ages ago, before session handlers existed, so it has now been removed, as session handlers are the new method for doing stuff like this.

the `delete_player()` function in the client has been fixed. The session handler should disconnect the player, not the client.